requirements.txt
# Dashboard & viz
streamlit
pandas
altair
matplotlib
wordcloud

# Scraper & DB
yt-dlp
psycopg2-binary
python-dotenv

# (Airflow, Redis, etc. only when you use them)
